python/2023/day19/test_day19.py
import pytest
import day19 as aoc


@pytest.fixture
def example1():
    puzzle_input = """px{a<2006:qkq,m>2090:A,rfg}
pv{a>1716:R,A}
lnx{m>1548:A,A}
rfg{s<537:gd,x>2440:R,A}
qs{s>3448:A,lnx}
qkq{x<1416:A,crn}
crn{x>2662:A,R}
in{s<1351:px,qqz}
qqz{s>2770:qs,m<1801:hdj,R}
gd{a>3333:R,R}
hdj{m>838:A,pv}

{x=787,m=2655,a=1222,s=2876}
{x=1679,m=44,a=2067,s=496}
{x=2036,m=264,a=79,s=2244}
{x=2461,m=1339,a=466,s=291}
{x=2127,m=1623,a=2188,s=1013}"""
    return aoc.parse(puzzle_input)


@pytest.fixture
def example2():
    puzzle_input = """   """
    return aoc.parse(puzzle_input)


@pytest.mark.skip(reason="Not implemented")
def test_parse_example1(example1):
    """Test that input is parsed properly."""
    assert example1 == ...


def test_part1_example1(example1):
    """Test part 1 on example input."""
    assert aoc.part1(example1) == 19114


def test_part2_example1(example1):
    """Test part 2 on example input."""
    assert aoc.part2(example1) == 167409079868000


def test_high_check_in():
    _range = [0, 4001]
    value = 1716
    assert aoc.check_in(value, _range, True) == [[1716, 4001], [0, 1717]]


def test_low_check_in():
    _range = [0, 4001]
    value = 1716
    assert aoc.check_in(value, _range, False) == [[0, 1716], [1715, 4001]]


@pytest.mark.skip(reason="Not implemented")
def test_part2_example2(example2):
    """Test part 2 on example input."""
    assert aoc.part2(example2) == ...
